pyproject.toml
[tool.poetry]
name = "fastapi-k8s"
version = "0.1.0"
description = ""
authors = ["svaponi <10941963+svaponi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
black = "^24.4.0"
pytest-httpserver = "^1.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
